+++ CHANGELOG.md
@@ -11,12 +11,11 @@
 - Added list of files with:
   - Status / Loading sign
   - Current transformer
-- Config is in hjson
 - Output file extensions are determined automatically
+- Dependencies can be specified for files
 
 ## TODO:
 - Can customize output file extensions
-- Dependencies can be specified for files
 - Add tests
 
 # 0.0.1-0.0.14